refactor client controllers, refactor clientController tests
