+++ src/main/resources/i18n/messages_cs.properties
@@ -43,7 +43,7 @@ label.client.cypherdetail.hintselect=Vybrat nápovědu
 label.client.cypherdetail.giveup=Vzdát šifru
 label.client.cypherdetail.giveupconfirm=Opravdu to chcete vzdát?
 label.client.finalplace.resultsannouncement=Vyhlášení výsledků
-label.client.finalplace.announcementalert=Vyhlášení proběhne dnes v 
+label.client.finalplace.announcementalert=Vyhlášení proběhne dnes v
 label.client.finalplace.address=Adresa
 label.client.finalplace.showonmap=Zobrazit na mapě
 label.client.hintlist.hint=Nápověda
