+++ src/main/resources/i18n/messages_en.properties
@@ -43,7 +43,7 @@ label.client.cypherdetail.hintselect=Select hint
 label.client.cypherdetail.giveup=Give up
 label.client.cypherdetail.giveupconfirm=Do you really want to give up?
 label.client.finalplace.resultsannouncement=Result announcement
-label.client.finalplace.announcementalert=Results will be announced today at 
+label.client.finalplace.announcementalert=Results will be announced today at
 label.client.finalplace.address=Address
 label.client.finalplace.showonmap=Show on map
 label.client.hintlist.hint=Hints
