+++ src/test/java/dk/cngroup/lentils/controller/client/HintControllerIntegrationTest.java
@@ -24,12 +24,14 @@ import static org.junit.Assert.assertNotEquals;
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = {LentilsApplication.class, DataConfig.class, ObjectGenerator.class})
 @Transactional
-public class ClientControllerTakeHintIntegrationTest extends AbstractClientControllerTest {
+public class HintControllerIntegrationTest extends AbstractClientControllerTest {
     private static final String HINT_NAME = "abcd";
     private static final int HINT_VALUE = 5;
 
     @Autowired
     private HintTakenService hintTakenService;
+    @Autowired
+    private HintController hintController;
 
     private Hint hint;
 
@@ -45,7 +47,7 @@ public class ClientControllerTakeHintIntegrationTest extends AbstractClientContr
     public void testTakeHintCorrectly() {
         setThatGameHasNotEnded();
 
-        final String returnValue = executeTestedMethod();
+        final String returnValue = executeGetHint();
 
         AssertionUtils.assertValueIsRedirection(returnValue);
         final int hintScore = hintTakenService.getHintScore(getFixtureTeam(), getFixtureCypher());
@@ -57,7 +59,7 @@ public class ClientControllerTakeHintIntegrationTest extends AbstractClientContr
         setThatGameHasNotEnded();
         setCypherStatusForTeam(getFixtureCypher(), CypherStatus.SOLVED, getFixtureTeam());
 
-        final String returnValue = executeTestedMethod();
+        final String returnValue = executeGetHint();
 
         AssertionUtils.assertValueIsRedirection(returnValue);
         final int hintScore = hintTakenService.getHintScore(getFixtureTeam(), getFixtureCypher());
@@ -68,15 +70,15 @@ public class ClientControllerTakeHintIntegrationTest extends AbstractClientContr
     public void testTakeHintAfterGameHasEnded() {
         setThatGameHasAlreadyEnded();
 
-        String returnValue = executeTestedMethod();
+        String returnValue = executeGetHint();
 
         AssertionUtils.assertValueIsRedirection(returnValue);
         final int hintScore = hintTakenService.getHintScore(getFixtureTeam(), getFixtureCypher());
         assertNotEquals(HINT_VALUE, hintScore);
     }
 
-    private String executeTestedMethod() {
-                return hintController.getHint(hint.getHintId(), getUserDetailsMock(), getFixtureCypher());
+    private String executeGetHint() {
+        return hintController.getHint(hint.getHintId(), getUserDetailsMock(), getFixtureCypher());
     }
 
     private Hint createTestHintForCypher(final Cypher cypher) {
